Preliminary Testing functionality
test_sandwiches allows you to specify your own thread ordering. test_kitchen_switching tries to do the same, but its a bit more complicated and uses some synchronization on the Carnivore and Vegetarian thread which is 1 not ideal, and 2 not fully working but may still be useful. Bug in my barrier implementation so I am using the pthread_barrier_t now and its working like a charm